Move locale logic to shared concern
